sell excess recs
